Linting

+++ .eslintrc.js
@@ -22,6 +22,6 @@ module.exports = {
     sourceType: 'module'
   },
   rules: {
-    quotes: ["error", "double"]
+    quotes: ["error", "single"]
   }
 }
